Move books state to higher layer
