+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Book = {
+    id: number,
+    title: string
+}
